+++ src/homework/tic_tac_toe/main.cpp
@@ -1,4 +1,5 @@
 #include "tic_tac_toe.h"
+#include "tic_tac_toe_manager.h"
 using namespace std;
 
 int main()
