+++ src/homework/tic_tac_toe/tic_tac_toe.h
@@ -29,23 +29,4 @@ class TicTacToe
         friend std::ostream& operator<<(std::ostream&, const TicTacToe&);
         friend std::istream& operator<<(std::istream&, TicTacToe&);
 };
-
-class TicTacToeManager
-{
-    private:
-        std::vector <TicTacToe> games;
-        void update_winner_count(std::string);
-
-        int x_wins = {0};
-        int o_wins = {0};
-        int ties = {0};
-
-    public:
-        void save_game(TicTacToe);
-        void get_winner_total(int&, int&, int&);
-
-        friend std::ostream& operator<<(std::ostream&, const TicTacToeManager&);
-
-
-};
 #endif
